# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Meer API
  version: 0.1.0
  description: REST API for the Meer app (mobile). Uses JWT bearer auth.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
  /auth/login:
    post:
      summary: Email/password login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Tokens and user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /dashboard/login:
    post:
      summary: Admin-only login for dashboard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Tokens and user (role will be ADMIN)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /dashboard/users/{id}:
    get:
      summary: Get a user by id (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: User
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
  /dashboard/support/contacts:
    get:
      summary: List contact messages (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 0 }
        - in: query
          name: pageSize
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: Paged contact messages
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponseSupportContact'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /dashboard/support/contacts/{id}:
    get:
      summary: Get contact message by id (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Contact message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupportContact'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Not found
    delete:
      summary: Delete contact message (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '204':
          description: Deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Not found
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: User not found
...
  /dashboard/stores:
    get:
      summary: Admin stores list
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 0 }
        - in: query
          name: pageSize
          schema: { type: integer, default: 20 }
        - in: query
          name: search
          schema: { type: string }
          description: Case-insensitive search across store name, tagline, description, neighborhood
        - in: query
          name: q
          schema: { type: string }
          description: Alias for search
        - in: query
          name: sort
          schema:
            type: string
            enum: [newest, oldest]
            default: newest
      responses:
        '200':
          description: Paged stores (admin)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponseDashboardStoreSummary'
  /dashboard/contents:
    get:
      summary: Admin contents list
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 0 }
        - in: query
          name: pageSize
          schema: { type: integer, default: 20 }
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: sort
          schema:
            type: string
            enum: [newest, oldest]
            default: newest
      responses:
        '200':
          description: Paged contents (admin)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponseGuideTop'
  /dashboard/users:
    get:
      summary: Admin users list
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 0 }
        - in: query
          name: pageSize
          schema: { type: integer, default: 20 }
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: sort
          schema:
            type: string
            enum: [newest, oldest]
            default: newest
      responses:
        '200':
          description: Paged users (admin)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponseAdminUser'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
  schemas:
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, format: password }
...
    AuthResponse:
      type: object
      properties:
        token: { type: string }
        refreshToken: { type: string }
        user:
          $ref: '#/components/schemas/Profile'
    Profile:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string, format: email }
        role: { type: string, enum: [USER, ADMIN] }
        avatarUrl: { type: string }
        bio: { type: string }
        notifyNewStores: { type: boolean }
        notifyPromos: { type: boolean }
        ownedThriftStore:
          $ref: '#/components/schemas/ThriftStore'
        createdAt: { type: string, format: date-time }
    StoreImage:
      type: object
      properties:
        id: { type: integer }
        url: { type: string }
        displayOrder: { type: integer }
        isCover: { type: boolean }
    SupportContact:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string, format: email }
        message: { type: string }
        createdAt: { type: string, format: date-time }
    PageResponseSupportContact:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/SupportContact' }
        page: { type: integer }
        hasNext: { type: boolean }
    ThriftStore:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        tagline: { type: string }
        coverImageUrl: { type: string }
        addressLine: { type: string }
        latitude: { type: number, format: double }
        longitude: { type: number, format: double }
        openingHours: { type: string }
        facebook: { type: string }
        instagram: { type: string }
        website: { type: string }
        phone: { type: string }
        whatsapp: { type: string }
        categories:
          type: array
          items: { type: string }
        rating: { type: number, format: double }
        reviewCount: { type: integer }
        distanceMeters: { type: number, format: double }
        walkTimeMinutes: { type: integer }
        neighborhood: { type: string }
        badgeLabel: { type: string }
        isFavorite: { type: boolean }
        description: { type: string }
        createdAt: { type: string, format: date-time }
        images:
          type: array
          items: { $ref: '#/components/schemas/StoreImage' }
...
    GuideTop:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: string }
        imageUrl: { type: string }
        thriftStoreId: { type: string, format: uuid }
        thriftStoreName: { type: string }
        createdAt: { type: string, format: date-time }
...
    DashboardStoreSummary:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        addressLine: { type: string }
        createdAt: { type: string, format: date-time }
    PageResponseDashboardStoreSummary:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/DashboardStoreSummary'
        page: { type: integer }
        hasNext: { type: boolean }
...
    PageResponseGuideTop:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/GuideTop'
        page: { type: integer }
        hasNext: { type: boolean }
    AdminUser:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        email: { type: string, format: email }
        role: { type: string, enum: [USER, ADMIN] }
        createdAt: { type: string, format: date-time }
    PageResponseAdminUser:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AdminUser'
        page: { type: integer }
        hasNext: { type: boolean }