# springboot/src/main/java/com/edufelip/meer/dto/DashboardStoreSummaryDto.java
package com.edufelip.meer.dto;

import java.time.Instant;
import java.util.UUID;

public record DashboardStoreSummaryDto(
        UUID id,
        String name,
        String addressLine,
        Instant createdAt
) {}