# springboot/src/main/java/com/edufelip/meer/dto/SupportContactDto.java
package com.edufelip.meer.dto;

import java.time.Instant;

public record SupportContactDto(
        Integer id,
        String name,
        String email,
        String message,
        Instant createdAt
) {}